+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.0+3
+
+* Bug-fix: bad RegExp in for private key
+* Bug-fix: Missing output directory in `google-play-version` command.
+* Bug-fix: Missing output directory in `app-store-version` command.
+
 ## 0.1.0+2
 
 * Re-generated the version.dart file, which is used to generate the version number.